requirements.txt
PyQt5==5.13.2
PyQt5-sip==12.7.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.2
pyserial==3.4
setuptools==39.1.0